+++ php/classes/integrations/elementor/widgets/elementor-recent-episodes-widget.php
@@ -3,6 +3,7 @@
 namespace SeriouslySimplePodcasting\Integrations\Elementor\Widgets;
 
 use SeriouslySimplePodcasting\Controllers\Episode_Controller;
+use SeriouslySimplePodcasting\Controllers\Episodes_Controller;
 
 class Elementor_Recent_Episodes_Widget extends \Elementor\Widget_Base {
 	public function get_name() {
@@ -23,7 +24,7 @@ class Elementor_Recent_Episodes_Widget extends \Elementor\Widget_Base {
 
 	protected function render() {
 		$settings          = $this->get_settings_for_display();
-		$episode_controller = new Episode_Controller( __FILE__, SSP_VERSION );
-		echo $episode_controller->render_recent_episodes();
+		$episodes_controller = new Episodes_Controller();
+		echo $episodes_controller->render_recent_episodes();
 	}
 }
